requirements.txt
altair
numpy
pandas
pydeck
streamlit
supabase-py
gotrue
postgrest-py
pytest
realtime-py
requests